add import alias column to dataclass table
