+++ experiment/src/org/labkey/experiment/api/ExpTableImpl.java
@@ -469,6 +469,25 @@ abstract public class ExpTableImpl<C extends Enum>
         addMethod("expObject", new ExpObjectTableMethodInfo(expObjectColumnName), Set.of());
     }
 
+    protected AliasedColumn createImportAliasColumn(String name, String prefix, String dataType)
+    {
+        AliasedColumn aliasedColumn = new AliasedColumn(this, name, _rootTable.getColumn("RowId"))
+        {
+            @Override
+            public boolean isNumericType()
+            {
+                // Issue 45374: don't apply number format to the RowId
+                return false;
+            }
+        };
+        aliasedColumn.setDisplayColumnFactory(new ImportAliasesDisplayColumnFactory(prefix));
+        aliasedColumn.setDescription("Display column for " + dataType + " import alias key/value pairs.");
+        aliasedColumn.setKeyField(false);
+        aliasedColumn.setRequired(false);
+        aliasedColumn.setHidden(true);
+        return aliasedColumn;
+    }
+
     private class ExpObjectTableMethodInfo extends AbstractTableMethodInfo
     {
         private final String _expObjectColumnName;
